+++ src/Toast/Toast.stories.tsx
@@ -1,7 +1,5 @@
-import React from "react";
+import React, { ComponentProps, useCallback } from "react";
 import { Story, Meta } from "@storybook/react/types-6-0";
-
-import { ToastContainerProps } from "react-toastify";
 
 import { Button } from "../Button";
 import { toast, Toast as T, ToastContainer, ToastProps } from "./index";
@@ -12,17 +10,48 @@ export default {
 } as Meta;
 
 export const Toast: Story<
-  ToastProps & Pick<ToastContainerProps, "position">
+  ToastProps &
+    Pick<
+      ComponentProps<typeof ToastContainer>,
+      "position" | "bordered" | "shadow"
+    >
 > = ({ type, title, message, copyError, ...args }) => {
+  const showLoadingToast = useCallback(() => {
+    const id = toast.loading(
+      "Loading",
+      "This is a loading toast, will change after 5 seconds"
+    );
+    setTimeout(() => {
+      toast.update(
+        id,
+        {
+          type: "success",
+          title: "Done",
+          message: "All loaded! Closing in 2 seconds",
+        },
+        { autoClose: 2000 }
+      );
+    }, 5000);
+  }, []);
+
   return (
-    <div className="section base-margin-top dbl-margin-bottom">
-      <h3 className="display-5">Toast</h3>
-      <Button.Primary onClick={() => toast(type, title, message, copyError)}>
-        Show toast
-      </Button.Primary>
+    <>
+      <div className="section base-margin-top dbl-margin-bottom">
+        <h3 className="display-5">Toast</h3>
+        <Button.Primary onClick={() => toast(type, title, message, copyError)}>
+          Show toast
+        </Button.Primary>
+
+        <p>Check source code to see how toasts called</p>
+      </div>
+      <div className="section base-margin-top dbl-margin-bottom">
+        <h3 className="display-5">Loding toast</h3>
+        <Button.Primary onClick={showLoadingToast}>
+          Show loading toast
+        </Button.Primary>
+      </div>
       <ToastContainer {...args} />
-      <p>Check source code to see how toasts called</p>
-    </div>
+    </>
   );
 };
 
@@ -31,7 +60,31 @@ Toast.parameters = {
     source: {
       code: `<Button.Primary onClick={() => toast(type, title, message, copyError)}>
   Show toast
-</Button.Primary>`,
+</Button.Primary>
+
+// Loading toast
+const showLoadingToast = useCallback(() => {
+  const id = toast.loading(
+    "Loading",
+    "This is a loading toast, will change after 5 seconds"
+  );
+  setTimeout(() => {
+    toast.update(
+      id,
+      {
+        type: "success",
+        title: "Done",
+        message: "All loaded! Closing in 2 seconds",
+      },
+      { autoClose: 2000 }
+    );
+  }, 5000);
+}, []);
+
+<Button.Primary onClick={showLoadingToast}>
+  Show loading toast
+</Button.Primary>
+`,
       language: "tsx",
     },
   },
@@ -42,4 +95,26 @@ Toast.args = {
   message: "Lorem ipsum dolor sit amet, consectetur adipiscing elit",
   type: "info",
   position: "bottom-right",
+  bordered: false,
+  shadow: "lg",
+};
+
+Toast.argTypes = {
+  position: {
+    options: [
+      "top-right",
+      "top-center",
+      "top-left",
+      "bottom-right",
+      "bottom-center",
+      "bottom-left",
+    ],
+    type: { name: "string", required: false },
+    control: "select",
+  },
+  shadow: {
+    options: ["lg", "md", "sm"],
+    type: { name: "string", required: false },
+    control: "radio",
+  },
 };