+++ src/Toast/ToastContainer.tsx
@@ -1,20 +1,31 @@
-import React, { FC } from "react";
+import React from "react";
 import {
   ToastContainer as ToastifyContainer,
   ToastContainerProps,
   Slide,
 } from "react-toastify";
+import { appendClass } from "src/utils";
 
-export const ToastContainer: FC<ToastContainerProps> = ({
+type ExtraPorps = {
+  bordered?: boolean;
+  shadow?: "sm" | "md" | "lg";
+};
+
+export const ToastContainer = ({
   position = "bottom-right",
   autoClose = 5000,
   draggable = false,
   hideProgressBar = true,
   containerId = "_GLOBAL_",
+  transition = Slide,
+  bordered,
+  shadow = "lg",
+  bodyClassName,
+  toastClassName,
   ...props
-}) => (
+}: ToastContainerProps & ExtraPorps): JSX.Element => (
   <ToastifyContainer
-    transition={Slide}
+    transition={transition}
     position={position}
     autoClose={autoClose}
     draggable={draggable}
@@ -22,6 +33,14 @@ export const ToastContainer: FC<ToastContainerProps> = ({
     containerId={containerId}
     {...props}
     closeButton={false}
+    bodyClassName={`${bodyClassName || ""}${appendClass(
+      bordered,
+      "toast--bordered"
+    )}`}
+    toastClassName={`${toastClassName || ""}${appendClass(
+      shadow === "md",
+      "toast--shadow-md"
+    )}${appendClass(shadow === "sm", "toast--shadow-sm")}`}
     style={{
       width: "unset",
     }}